Fixing a SEGFAULT for unaccessable files
I've run into this issue of having *SIGSEGV* while a directory that I was reading
wasn't fully checked out by `dep` (usual behavior). That caused a dangling symlink
in that directory. And, that dangling symlink produced hard to debug SIGSEGV.

Proper behavior would be to return an error that states a given file isn't readable.