Introduce 3-way strategic merging for all but ConfigMaps
The strategic merge test is from the k8s docs:
https://kubernetes.io/docs/tasks/manage-kubernetes-objects/declarative-config/#merge-patch-calculation

We factored all the patching logic out into its own package, since it
introduced a lot of noise in the primary source file.

I might've gotten a little happy with struct composition. Could be
clearer, I think.

We also give an indication of which fields are triggering the update
in the log. That should help debug superfluous updates and speed up
installation considerably.