Verify predicate mutation until we decide it's bad
We're all adults here! If we do decide it's bad, I think we should
stop returning possible errors from Transform, mostly to support
method chaining. But if we do that, we need to encourage creating
Transformers safely, i.e. returning closures after dealing with
potential errors instead of dealing with them in the Transformer.