Using Replace instead of Strategic option
This forces an update of the original spec rather than a JSON merge
patch, which is effectively identical to strategic patches for
ConfigMaps anyway, i.e. no lists to deal with.

We still need JSON merge patch creation, but only for types not
registered in the schema, so no need to have an option to select which
merge patch.

Also fixed some of the spec mutation for creates.