Added a `Replace` option and refactored patch logic for clarity (#25)
* Refactored patch logic to add clarity and remove redundancy

* No more special case updates of ConfigMaps

* Introduced `Replace` option

This forces an update of the original spec rather than a JSON merge
patch, which is effectively identical to strategic patches for
ConfigMaps anyway, i.e. no lists to deal with.

We still need JSON merge patch creation, but only for types not
registered in the schema, so no need to have an option to select which
merge patch.

Also fixed some of the spec mutation for creates.

* Separate lastApplied calculation from patch creation

This greatly simplifies the patch logic, enabling me to replace the
struct embedding with a single 'if' clause. Lets clients deal with
nil return values instead of calling Empty(), too. I think it's
clearer overall.