When update fails, retry an overwrite
This involves overlaying the manifest fields onto the current
resource, which might contain server-provided, immutable fields absent
from the manifest, e.g. a service's clusterIP or a creationTimestamp.

We take a conservative approach to overlaying objects with entirely
different structures, as might be the case when updating a new
release. This may still error, but there's not a lot we can do other
than delete/create. But I'd prefer the user to make that call.