Include the manifestival annotation in the last-applied annotation
We're applying it, so it ought to be in the last-applied annotation,
right?

Additionally, we must update the manifestival metadata in the dry-run
logic so that it's not a part of the diff.

And while we were in there we renamed the variable original->current
to more accurately represent what's going on... the original spec is
taken from the current spec in patch.New