Prevent mutation by Filter
Fixes #75

The ability to chain `Apply` calls to the result of `Filter` made it
necessary to return the actual manifest resources instead of deep
copies. So now the Predicates are passed deep copies, making it
impossible to alter resources by filtering.

Only the Transform method can alter resources, and its multi-value
return prevents the chaining of `Apply` calls.