Make sure the 'manifestival' annotation is kept regardless of manifest state
Currently, the manifestival annotation relies on the annotation being
added to the manifest itself (the object in the manifest gets annotated,
not its deep-copy) and then carried through reusing that same manifest
instance all-the-time.

If a user ever generates a new manifest, even if it is byte-by-byte
identical, the manifestival annotation gets patched away as the new
manifest won't carry that implicit state.

This fixes that by carrying over the manifestival annotation regardless
of state and patch status, if it was set before applying the patch.